requirement.txt
diffusers
torch
accelerate
fastapi
pillow
pandas
transformers
openpyxl